Restarting containers on ConfigMap updates, using livenessProbes
The basic idea is to use a livenessProbe to restart a container when
files in a specified config
directory change (f.e. because of a change in a ConfigMap attached as a
volume).

app.py reads a 'MAGIC' value from the /etc/app/config.yml file at the
start. It is then available via the http server at \<ip>:8000/var.

config_update_check.py checks the modification time of a specified
folder and stores it in a temporary file.
If the registered value is larger then the previous one, script fails.
By using this script in a
livenessProbe defined for the app container, we can trigger restarts of
the container when files in
the config directory change.